vault router push works!!:)

+++ keeper/src/pages/VaultPage.vue
@@ -1,9 +1,9 @@
 <template>
   <div class="vaultPage row w-100 justify-content-center">
     <div class="col">
-      <router-link :to="{ name: 'Account' }" class="nav-link">
-        <i class="fa fa-trash text-danger" @click="deleteVault" v-if="state.vault.creatorId == state.account.id" aria-hidden="true"></i>
-      </router-link>
+      <!-- <router-link :to="{ name: 'Account' }" class="nav-link"> -->
+      <i class="fa fa-trash text-danger" @click="deleteVault" v-if="state.vault.creatorId == state.account.id" aria-hidden="true"></i>
+      <!-- </router-link> -->
       {{ state.vault.name }}
     </div>
     <keep-component v-for="keep in state.keeps" :key="keep.id" :keep-prop="keep" />
@@ -15,12 +15,14 @@ import { reactive } from '@vue/reactivity'
 import { computed, onMounted } from '@vue/runtime-core'
 import { AppState } from '../AppState'
 import { keepService } from '../services/KeepService'
-import { useRoute } from 'vue-router'
+import { useRoute, useRouter } from 'vue-router'
 import { vaultService } from '../services/VaultService'
+import swal from 'sweetalert2'
 export default {
   name: 'VaultPage',
   setup() {
     const route = useRoute()
+    const router = useRouter()
     const state = reactive({
       keeps: computed(() => AppState.keeps),
       vault: computed(() => AppState.activeVault),
@@ -31,7 +33,23 @@ export default {
     return {
       state,
       deleteVault() {
-        vaultService.deleteVault(state.vault.id)
+        swal.fire({
+          title: 'Are you sure you want to delete this vault?',
+          icon: 'warning',
+          buttons: true,
+          dangerMode: true
+        })
+          .then((willDelete) => {
+            if (willDelete.isConfirmed) {
+              swal.fire({
+                icon: 'success'
+              })
+              vaultService.deleteVault(state.vault.id)
+              router.push({ name: 'Account' })
+            } else {
+              swal.fire('Cancelled')
+            }
+          })
       }
     }
   },
